+++ flutter/lib/benchmark/benchmark.dart
@@ -7,6 +7,7 @@ import 'package:mlperfbench/backend/bridge/run_settings.dart';
 import 'package:mlperfbench/device_info.dart';
 import 'package:mlperfbench/protos/backend_setting.pb.dart' as pb;
 import 'package:mlperfbench/protos/mlperf_task.pb.dart' as pb;
+import 'package:mlperfbench/resources/resource.dart';
 import 'package:mlperfbench/resources/resource_manager.dart';
 import 'info.dart';
 import 'run_mode.dart';
